+++ monorepo/src/espu/bezier/exceptions.py
@@ -1,0 +1,4 @@
+from espu.core import CoreError
+
+class CurveNotBakedError(CoreError):
+    """Raised when attempting to resolve_uniform() a curve before it has been baked."""